# Optimize `ParseErrors::transpose` method